make stars look less jarring
when the site loads, it makes a request to get the current number of stars for the project and replaces it with the number. before this (or if it doesn't succeed), "starsHere" is shown. IMO, this looks very jarring; so, instead, we just show the star icon and nothing else prior to the request finishing, which IMO looks much more natural.